Change to long() to enable training reproduction for the nar model 3

+++ utils/nar_beam_search.py
@@ -118,5 +118,5 @@ class Beamsearch(object):
         hyp = -1 * torch.ones(self.batch_size, self.num_nodes, dtype=torch.long).to(self.device)
         for j in range(len(self.prev_Ks) - 1, -2, -1):
             hyp[:, j + 1] = self.next_nodes[j + 1].gather(1, k.long()).view(1, self.batch_size)
-            k = self.prev_Ks[j].gather(1, k)
+            k = self.prev_Ks[j].gather(1, k.long())
         return hyp
